window dataset unit test

+++ tests/image_reader_test.py
@@ -354,7 +354,7 @@ class ImageReaderTest(tf.test.TestCase):
         # Check data type
         self.assertGreaterEqual(image.min(), 0)
         self.assertLessEqual(image.max(), 255)
-        self.assertEqual(image.dtype, np.int32)
+        self.assertEqual(image.dtype, np.float32)
         self.assertEqual(np.unique(image).size, 2)
         # Check shape
         self.assertEqual(image.ndim, 5)
